🐞 fix: psutil.cmdline cause privelege issue on macos
when psutil.cmd_line is called, it'll raise exception. So I have to
save process info into redis cache.

Beside, there's sanic multi-processing issue:
```
objc[96314]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called.
objc[96315]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called.
objc[96314]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called. We cannot safely call it or ignore it in the fork() child process. Crashing instead. Set a breakpoint on objc_initializeAfterForkError to debug.
objc[96315]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called. We cannot safely call it or ignore it in the fork() child process. Crashing instead. Set a breakpoint on objc_initializeAfterForkError to debug.
```
start sanic in just one worker will solve the problem. so it's a sanic's
bug.